rename synctexParser.ts to synctexjs.ts
