+++ src/components/locator.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode'
 import * as fs from 'fs'
 import * as path from 'path'
 import * as cp from 'child_process'
-import * as syncTexJs from './synctexParser'
+import * as syncTexJs from './synctexjs'
 
 import {Extension} from '../main'
