+++ src/components/synctexjs.ts
